requirements.txt
anthropic>=0.40.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pypdf>=5.0.0
python-dotenv>=1.0.0
gradio>=5.0.0
pyyaml>=6.0
rich>=13.7.0
